Minor refactoring and add Licenses
